Publish beta packages to NPM

+++ source/swap-erc20/contracts/interfaces/INoReturnERC20.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.23;
 
 interface INoReturnERC20 {
-  function transferFrom(address from, address to, uint value) external;
+  function transferFrom(address from, address to, uint256 value) external;
 }
